requirements.txt
torch==1.7.1+cu110
torchvision==0.8.2+cu110
numpy==1.21.1
opencv-python==4.1.1.26
h5py==2.10.0
Cython==0.29
Pillow==8.3.1
scipy==1.7.0
easydict==1.9